requirements.txt
pytest==3.10.1
Werkzeug==0.11.15
Flask_Cors==3.0.10
requests==2.21.0
Flask==0.12.1
PyJWT==2.0.1
